2.0.0.0-dev70 * Fixed bugs:   * Fixed an issue where the schedule of recurring payments was not displayed in the shopping cart   * Fixed an issue with displaying tax class names in the Customer Groups grid   * Fixed an issue with testing Solr connection   * Fixed an issue with using custom module front name   * Fixed an issue with USPS and DHL usage in the production mode * Modularity improvements:   * Consolidated all logic related to Layered Navigation in one separate module * Framework Improvements:   * Covered Magento library components with unit tests:      * Magento/Interception      * Magento/ObjectManager      * Magento/Message      * Magento/Module      * Magento/Mail      * Magento/Object      * Magento/Math * Updated XML files to include a reference to the schema file in a form of a relative path * Updated code to be PSR-2 compliant

+++ app/code/Magento/Catalog/Controller/Adminhtml/Product/Initialization/StockDataFilter.php
@@ -21,7 +21,6 @@
  * @copyright   Copyright (c) 2014 X.commerce, Inc. (http://www.magentocommerce.com)
  * @license     http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-
 namespace Magento\Catalog\Controller\Adminhtml\Product\Initialization;
 
 class StockDataFilter
@@ -57,10 +56,11 @@ class StockDataFilter
         }
 
         if ($stockData['use_config_manage_stock'] == 1 && !isset($stockData['manage_stock'])) {
-            $stockData['manage_stock'] = $this->storeManager->getStore()
-                ->getConfig(\Magento\CatalogInventory\Model\Stock\Item::XML_PATH_MANAGE_STOCK);
+            $stockData['manage_stock'] = $this->storeManager->getStore()->getConfig(
+                \Magento\CatalogInventory\Model\Stock\Item::XML_PATH_MANAGE_STOCK
+            );
         }
-        if (isset($stockData['qty']) && (float)$stockData['qty'] > self::MAX_QTY_VALUE) {
+        if (isset($stockData['qty']) && (double)$stockData['qty'] > self::MAX_QTY_VALUE) {
             $stockData['qty'] = self::MAX_QTY_VALUE;
         }
 
@@ -74,4 +74,4 @@ class StockDataFilter
 
         return $stockData;
     }
-} 
+}
