+++ app/code/Magento/Paypal/Model/Resource/Report/Settlement.php
@@ -73,7 +73,7 @@ class Settlement extends \Magento\Core\Model\Resource\Db\AbstractDb
     {
         $rows = $object->getRows();
         if (is_array($rows)) {
-            $adapter  = $this->_getWriteAdapter();
+            $adapter = $this->_getWriteAdapter();
             $reportId = (int)$object->getId();
             try {
                 $adapter->beginTransaction();
@@ -86,15 +86,19 @@ class Settlement extends \Magento\Core\Model\Resource\Db\AbstractDb
                      * Converting dates
                      */
                     $completionDate = new \Magento\Stdlib\DateTime\Date($rows[$key]['transaction_completion_date']);
-                    $rows[$key]['transaction_completion_date'] = $this->_coreDate
-                        ->date(null, $completionDate->getTimestamp());
+                    $rows[$key]['transaction_completion_date'] = $this->_coreDate->date(
+                        null,
+                        $completionDate->getTimestamp()
+                    );
                     $initiationDate = new \Magento\Stdlib\DateTime\Date($rows[$key]['transaction_initiation_date']);
-                    $rows[$key]['transaction_initiation_date'] = $this->_coreDate
-                        ->date(null, $initiationDate->getTimestamp());
+                    $rows[$key]['transaction_initiation_date'] = $this->_coreDate->date(
+                        null,
+                        $initiationDate->getTimestamp()
+                    );
                     /*
                      * Converting numeric
                      */
-                    $rows[$key]['fee_amount'] = (float)$rows[$key]['fee_amount'];
+                    $rows[$key]['fee_amount'] = (double)$rows[$key]['fee_amount'];
                     /*
                      * Setting reportId
                      */
@@ -123,10 +127,13 @@ class Settlement extends \Magento\Core\Model\Resource\Db\AbstractDb
     public function loadByAccountAndDate(\Magento\Paypal\Model\Report\Settlement $report, $accountId, $reportDate)
     {
         $adapter = $this->_getReadAdapter();
-        $select  = $adapter->select()
-            ->from($this->getMainTable())
-            ->where('account_id = :account_id')
-            ->where('report_date = :report_date');
+        $select = $adapter->select()->from(
+            $this->getMainTable()
+        )->where(
+            'account_id = :account_id'
+        )->where(
+            'report_date = :report_date'
+        );
 
         $data = $adapter->fetchRow($select, array(':account_id' => $accountId, ':report_date' => $reportDate));
         if ($data) {
